// src/common/enums/certificate-message.enum.ts
export enum CertificateMessages {
    NOT_FOUND = 'Certificate not found',
    ALREADY_EXISTS = 'Certificate already exists',
    CREATED = 'Certificate created successfully',
    UPDATED = 'Certificate updated successfully',
    DELETED = 'Certificate deleted successfully',

    UNAUTHORIZED_ACCESS = 'Unauthorized access',
    INVALID_CREDENTIALS = 'Invalid credentials',
}